test(evm): ante handler setup ctx coverage
